Move run to its own service layer
