+++ src/main/java/marquez/MarquezContext.java
@@ -30,6 +30,7 @@ import marquez.db.TagDao;
 import marquez.service.DatasetService;
 import marquez.service.JobService;
 import marquez.service.NamespaceService;
+import marquez.service.RunService;
 import marquez.service.RunTransitionListener;
 import marquez.service.SourceService;
 import marquez.service.TagService;
@@ -70,6 +71,7 @@ public final class MarquezContext {
 
   @Getter private final ImmutableList<Object> resources;
   @Getter private final JdbiExceptionExceptionMapper jdbiException;
+  @Getter private final RunService runService;
 
   private MarquezContext(
       @NonNull final Jdbi jdbi,
@@ -98,18 +100,19 @@ public final class MarquezContext {
     this.datasetService =
         new DatasetService(
             namespaceDao, sourceDao, datasetDao, datasetFieldDao, datasetVersionDao, tagDao);
+    this.runService =
+        new RunService(
+            jobVersionDao,
+            datasetDao,
+            runArgsDao,
+            runDao,
+            datasetVersionDao,
+            runStateDao,
+            Lists.newArrayList(runTransitionListeners));
+
     this.jobService =
         new JobService(
-            namespaceDao,
-            datasetDao,
-            datasetVersionDao,
-            jobDao,
-            jobVersionDao,
-            jobContextDao,
-            runDao,
-            runArgsDao,
-            runStateDao,
-            runTransitionListeners);
+            namespaceDao, datasetDao, jobDao, jobVersionDao, jobContextDao, runDao, runService);
     this.tagService = new TagService(tagDao);
     this.tagService.init(tags);
     this.serviceExceptionMapper = new MarquezServiceExceptionMapper();
@@ -118,8 +121,8 @@ public final class MarquezContext {
     this.namespaceResource = new NamespaceResource(namespaceService);
     this.sourceResource = new SourceResource(sourceService);
     this.datasetResource =
-        new DatasetResource(namespaceService, datasetService, jobService, tagService);
-    this.jobResource = new JobResource(namespaceService, jobService);
+        new DatasetResource(namespaceService, datasetService, tagService, runService);
+    this.jobResource = new JobResource(namespaceService, jobService, runService);
     this.tagResource = new TagResource(tagService);
 
     this.resources =
