+++ src/test/java/marquez/api/JobResourceTest.java
@@ -42,6 +42,7 @@ import marquez.common.models.NamespaceName;
 import marquez.common.models.RunId;
 import marquez.service.JobService;
 import marquez.service.NamespaceService;
+import marquez.service.RunService;
 import marquez.service.exceptions.MarquezServiceException;
 import marquez.service.models.Job;
 import marquez.service.models.JobMeta;
@@ -77,11 +78,12 @@ public class JobResourceTest {
 
   @Mock private NamespaceService namespaceService;
   @Mock private JobService jobService;
+  @Mock private RunService runService;
   private JobResource jobResource;
 
   @Before
   public void setUp() {
-    jobResource = spy(new JobResource(namespaceService, jobService));
+    jobResource = spy(new JobResource(namespaceService, jobService, runService));
   }
 
   @Test
@@ -158,7 +160,7 @@ public class JobResourceTest {
 
     when(namespaceService.exists(NAMESPACE_NAME)).thenReturn(true);
     when(jobService.exists(NAMESPACE_NAME, JOB_NAME)).thenReturn(true);
-    when(jobService.createRun(NAMESPACE_NAME, JOB_NAME, runMeta)).thenReturn(run);
+    when(runService.createRun(NAMESPACE_NAME, JOB_NAME, runMeta)).thenReturn(run);
 
     doReturn(runLocation).when(jobResource).locationFor(uriInfo, run);
 
@@ -179,7 +181,7 @@ public class JobResourceTest {
 
     when(namespaceService.exists(NAMESPACE_NAME)).thenReturn(true);
     when(jobService.exists(NAMESPACE_NAME, JOB_NAME)).thenReturn(true);
-    when(jobService.createRun(NAMESPACE_NAME, JOB_NAME, runMeta)).thenReturn(run);
+    when(runService.createRun(NAMESPACE_NAME, JOB_NAME, runMeta)).thenReturn(run);
 
     doReturn(runLocation).when(jobResource).locationFor(uriInfo, run);
 
@@ -197,7 +199,7 @@ public class JobResourceTest {
 
     when(namespaceService.exists(NAMESPACE_NAME)).thenReturn(true);
     when(jobService.exists(NAMESPACE_NAME, JOB_NAME)).thenReturn(true);
-    when(jobService.runExists(runIdExists)).thenReturn(true);
+    when(runService.runExists(runIdExists)).thenReturn(true);
 
     assertThatExceptionOfType(RunAlreadyExistsException.class)
         .isThrownBy(
@@ -234,7 +236,7 @@ public class JobResourceTest {
   public void testGetRun() throws MarquezServiceException {
     final Run run = newRunWith(RUN_ID);
 
-    when(jobService.getRun(RUN_ID)).thenReturn(Optional.of(run));
+    when(runService.getRun(RUN_ID)).thenReturn(Optional.of(run));
 
     final Response response = jobResource.getRun(RUN_ID);
     assertThat(response.getStatus()).isEqualTo(200);
@@ -243,7 +245,7 @@ public class JobResourceTest {
 
   @Test
   public void testGetRun_notFound() throws MarquezServiceException {
-    when(jobService.getRun(RUN_ID)).thenReturn(Optional.empty());
+    when(runService.getRun(RUN_ID)).thenReturn(Optional.empty());
 
     assertThatExceptionOfType(RunNotFoundException.class)
         .isThrownBy(() -> jobResource.getRun(RUN_ID))
@@ -254,7 +256,7 @@ public class JobResourceTest {
   public void testListRuns() throws MarquezServiceException {
     when(namespaceService.exists(NAMESPACE_NAME)).thenReturn(true);
     when(jobService.exists(NAMESPACE_NAME, JOB_NAME)).thenReturn(true);
-    when(jobService.getAllRunsFor(NAMESPACE_NAME, JOB_NAME, 4, 0)).thenReturn(RUNS);
+    when(runService.getAllRunsFor(NAMESPACE_NAME, JOB_NAME, 4, 0)).thenReturn(RUNS);
 
     final Response response = jobResource.listRuns(NAMESPACE_NAME, JOB_NAME, 4, 0);
     assertThat(response.getStatus()).isEqualTo(200);
@@ -265,7 +267,7 @@ public class JobResourceTest {
   public void testListRuns_empty() throws MarquezServiceException {
     when(namespaceService.exists(NAMESPACE_NAME)).thenReturn(true);
     when(jobService.exists(NAMESPACE_NAME, JOB_NAME)).thenReturn(true);
-    when(jobService.getAllRunsFor(NAMESPACE_NAME, JOB_NAME, 4, 0)).thenReturn(ImmutableList.of());
+    when(runService.getAllRunsFor(NAMESPACE_NAME, JOB_NAME, 4, 0)).thenReturn(ImmutableList.of());
 
     final Response response = jobResource.listRuns(NAMESPACE_NAME, JOB_NAME, 4, 0);
     assertThat(response.getStatus()).isEqualTo(200);
@@ -274,7 +276,7 @@ public class JobResourceTest {
 
   @Test
   public void testMarkRunAs_throwOnIdNotFound() throws MarquezServiceException {
-    when(jobService.runExists(RUN_ID)).thenReturn(false);
+    when(runService.runExists(RUN_ID)).thenReturn(false);
 
     assertThatExceptionOfType(RunNotFoundException.class)
         .isThrownBy(() -> jobResource.markRunAs(RUN_ID, RUNNING, TRANSITIONED_AT))
@@ -283,7 +285,7 @@ public class JobResourceTest {
 
   @Test
   public void testMarkRunAsRunning() throws MarquezServiceException {
-    when(jobService.runExists(RUN_ID)).thenReturn(true);
+    when(runService.runExists(RUN_ID)).thenReturn(true);
 
     final Run running = newRunWith(RUN_ID, RUNNING, TRANSITIONED_AT);
     doReturn(Response.ok(running).build()).when(jobResource).getRun(RUN_ID);
@@ -295,7 +297,7 @@ public class JobResourceTest {
 
   @Test
   public void testMarkRunAsCompleted() throws MarquezServiceException {
-    when(jobService.runExists(RUN_ID)).thenReturn(true);
+    when(runService.runExists(RUN_ID)).thenReturn(true);
 
     final Run completed = newRunWith(RUN_ID, COMPLETED, TRANSITIONED_AT);
     doReturn(Response.ok(completed).build()).when(jobResource).getRun(RUN_ID);
@@ -307,7 +309,7 @@ public class JobResourceTest {
 
   @Test
   public void testMarkRunAsFailed() throws MarquezServiceException {
-    when(jobService.runExists(RUN_ID)).thenReturn(true);
+    when(runService.runExists(RUN_ID)).thenReturn(true);
 
     final Run failed = newRunWith(RUN_ID, FAILED, TRANSITIONED_AT);
     doReturn(Response.ok(failed).build()).when(jobResource).getRun(RUN_ID);
@@ -319,7 +321,7 @@ public class JobResourceTest {
 
   @Test
   public void testMarkRunAsAborted() throws MarquezServiceException {
-    when(jobService.runExists(RUN_ID)).thenReturn(true);
+    when(runService.runExists(RUN_ID)).thenReturn(true);
 
     final Run aborted = newRunWith(RUN_ID, ABORTED, TRANSITIONED_AT);
     doReturn(Response.ok(aborted).build()).when(jobResource).getRun(RUN_ID);
